Splitting sandpack from main bundle (#4256)
* Initial Commit

* play with it

* oops

* easier repro

* import type

* remove `suspense: true`

* Add patch for next

* Patch package

* Add fallback

* Enable flag

* Fixes local dev env and adds better fallback for codeblock

* Adds fallback for sandpack (should work fine)

* turn off concurrentFeatures

* Revert "turn off concurrentFeatures"

This reverts commit 50158ecbd33969e707a2a91a54e822e90c2ebfde.

* Update SandpackWrapper.tsx

* Removed flags and setTimeouts

* add timeouts and promise again

* Adds bottom bezel and scroll to sandpack fallback

* tinker bottombezel and remove console

* Update CodeBlock.tsx

* Update SandpackWrapper.tsx

* removing overflows to avoid explicit scrolls

* upgrade nextjs to canary

* Rm patch

* Fix TS

* Bump Next

* No more CSS jumping

* Reverts the canary to use the latest Next.js `12.0.10`
