+++ beta/next.config.js
@@ -11,7 +11,7 @@ module.exports = {
   experimental: {
     plugins: true,
     // TODO: this doesn't work because https://github.com/vercel/next.js/issues/30714
-    // concurrentFeatures: true,
+    concurrentFeatures: false,
     scrollRestoration: true,
   },
   async redirects() {
@@ -27,7 +27,7 @@ module.exports = {
   },
   webpack: (config, {dev, isServer, ...options}) => {
     if (process.env.ANALYZE) {
-      const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer')
+      const {BundleAnalyzerPlugin} = require('webpack-bundle-analyzer');
       config.plugins.push(
         new BundleAnalyzerPlugin({
           analyzerMode: 'static',
