--- beta/src/components/MDX/CodeBlock/CodeBlock.module.css
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) Facebook, Inc. and its affiliates.
- */
-
-.codeViewer {
-  padding: 6px 0.5rem !important;
-}
