+++ beta/src/pages/_document.tsx
@@ -5,16 +5,15 @@
 import * as React from 'react';
 import Document, {Html, Head, Main, NextScript} from 'next/document';
 
-class MyDocument extends Document {
-  render() {
-    //  @todo specify language in HTML?
-    return (
-      <Html lang="en">
-        <Head />
-        <body className="font-sans antialiased text-lg bg-wash dark:bg-wash-dark text-secondary dark:text-secondary-dark leading-base">
-          <script
-            dangerouslySetInnerHTML={{
-              __html: `
+const MyDocument = () => {
+  //  @todo specify language in HTML?
+  return (
+    <Html lang="en">
+      <Head />
+      <body className="font-sans antialiased text-lg bg-wash dark:bg-wash-dark text-secondary dark:text-secondary-dark leading-base">
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `
                 (function () {
                   function setTheme(newTheme) {
                     window.__theme = newTheme;
@@ -53,14 +52,13 @@ class MyDocument extends Document {
                   });
                 })();
               `,
-            }}
-          />
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
+          }}
+        />
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+};
 
 export default MyDocument;
